Better error message for a route name typo?
Hi!
Running this test file shows error:
  [Sat Sep 22 07:28:16 2018] [debug] Using default_handler to render data since 'openapi' was not found in stash. Set 'handler' in stash to avoid this message.

It took a bit of time to figure out what this meant.

If you take a look at the openapi.yaml at the bottom,
you find that there is a typo in the path name:

    x-mojo-name: getEcho,

I managed to leave the comma there, when refactoring an API spec from json to yaml.
Essentially there is no such method/route to handle the GET /echo -endpoint.

I'd prefer to have a solid error message saying that there is no implementation for the
endpoint, instead of complaining about the missing openapi renderer.

Later on there is a 501 error Not Implemented, which is rather spot on.

Would it be possible to have some static safety when loading the OpenAPI-plugin
to check that all the endpoints have an implementation, or atleast a matching route?